+++ crates/unftp-auth-jsonfile/tests/main.rs
@@ -1,6 +1,6 @@
 #![allow(missing_docs)]
 
-use libunftp::auth::{Authenticator, DefaultUser};
+use libunftp::auth::Authenticator;
 use std::path::PathBuf;
 use unftp_auth_jsonfile::JsonFileAuthenticator;
 
@@ -17,7 +17,7 @@ async fn credentials_from_file_type_plain() {
     let path = input_file_path("cred.json".to_string());
 
     let json_auther = JsonFileAuthenticator::from_file(path).unwrap();
-    assert_eq!(json_auther.authenticate("testuser", &"testpassword".into()).await.unwrap(), DefaultUser);
+    assert_eq!(json_auther.authenticate("testuser", &"testpassword".into()).await.unwrap().username, "testuser");
 }
 
 #[tokio::test(flavor = "current_thread")]
@@ -25,7 +25,7 @@ async fn credentials_from_file_type_gzipped() {
     let path = input_file_path("cred.json.gz".to_string());
 
     let json_auther = JsonFileAuthenticator::from_file(path).unwrap();
-    assert_eq!(json_auther.authenticate("testuser", &"testpassword".into()).await.unwrap(), DefaultUser);
+    assert_eq!(json_auther.authenticate("testuser", &"testpassword".into()).await.unwrap().username, "testuser");
 }
 
 #[tokio::test(flavor = "current_thread")]
@@ -33,5 +33,5 @@ async fn credentials_from_file_type_gzipped_base64() {
     let path = input_file_path("cred.json.gz.b64".to_string());
 
     let json_auther = JsonFileAuthenticator::from_file(path).unwrap();
-    assert_eq!(json_auther.authenticate("testuser", &"testpassword".into()).await.unwrap(), DefaultUser);
+    assert_eq!(json_auther.authenticate("testuser", &"testpassword".into()).await.unwrap().username, "testuser");
 }