+++ src/auth/mod.rs
@@ -11,10 +11,10 @@
 //! async-trait = "0.1.50"
 //! ```
 //!
-//! 2. Implement the [`Authenticator`] trait and optionally the [`UserDetail`] trait:
+//! 2. Implement the [`Authenticator`] trait and optionally the [`UserDetail`] and [`UserDetailProvider`] traits:
 //!
 //! ```no_run
-//! use libunftp::auth::{Authenticator, AuthenticationError, UserDetail, Credentials};
+//! use libunftp::auth::{Authenticator, AuthenticationError, Principal, UserDetail, UserDetailProvider, UserDetailError, Credentials};
 //! use async_trait::async_trait;
 //! use unftp_sbe_fs::Filesystem;
 //!
@@ -22,9 +22,9 @@
 //! struct RandomAuthenticator;
 //!
 //! #[async_trait]
-//! impl Authenticator<RandomUser> for RandomAuthenticator {
-//!     async fn authenticate(&self, _username: &str, _creds: &Credentials) -> Result<RandomUser, AuthenticationError> {
-//!         Ok(RandomUser{})
+//! impl Authenticator for RandomAuthenticator {
+//!     async fn authenticate(&self, _username: &str, _creds: &Credentials) -> Result<Principal, AuthenticationError> {
+//!         Ok(Principal { username: _username.to_string() })
 //!     }
 //! }
 //!
@@ -38,6 +38,18 @@
 //!         write!(f, "RandomUser")
 //!     }
 //! }
+//!
+//! #[derive(Debug)]
+//! struct RandomUserDetailProvider;
+//!
+//! #[async_trait]
+//! impl UserDetailProvider for RandomUserDetailProvider {
+//!     type User = RandomUser;
+//!
+//!     async fn provide_user_detail(&self, _principal: &Principal) -> Result<RandomUser, UserDetailError> {
+//!         Ok(RandomUser {})
+//!     }
+//! }
 //! ```
 //!
 //! 3. Initialize it with the server:
@@ -45,10 +57,15 @@
 //! ```
 //! # // Make it compile
 //! # type RandomAuthenticator = libunftp::auth::AnonymousAuthenticator;
-//! let server = libunftp::Server::with_authenticator(
+//! # type RandomUserDetailProvider = libunftp::auth::DefaultUserDetailProvider;
+//! # type RandomUser = libunftp::auth::DefaultUser;
+//! let server = libunftp::ServerBuilder::with_authenticator(
 //!   Box::new(move || { unftp_sbe_fs::Filesystem::new("/srv/ftp").unwrap() }),
 //!   std::sync::Arc::new(RandomAuthenticator{})
-//! );
+//! )
+//! .user_detail_provider(std::sync::Arc::new(libunftp::auth::DefaultUserDetailProvider))
+//! .build()
+//! .unwrap();
 //! ```
 //!
 //! [`Server`]: ../struct.Server.html
@@ -63,4 +80,7 @@ pub(crate) mod authenticator;
 pub use authenticator::{AuthenticationError, Authenticator, ClientCert, Credentials, Principal};
 
 mod user;
-pub use user::{DefaultUser, UserDetail, UserDetailError, UserDetailProvider};
+pub use user::{DefaultUser, DefaultUserDetailProvider, UserDetail, UserDetailError, UserDetailProvider};
+
+mod pipeline;
+pub use pipeline::AuthenticationPipeline;